+++ atrium-api/src/types/string.rs
@@ -413,6 +413,76 @@ impl Serialize for Language {
     }
 }
 
+/// A [Timestamp Identifier].
+///
+/// [Timestamp Identifier]: https://atproto.com/specs/record-key#record-key-type-tid
+#[derive(Clone, Debug, PartialEq, Eq, Serialize)]
+#[serde(transparent)]
+pub struct Tid(String);
+string_newtype!(Tid);
+
+impl Tid {
+    #[allow(
+        clippy::borrow_interior_mutable_const,
+        clippy::declare_interior_mutable_const
+    )]
+    /// Parses a `TID` from the given string.
+    pub fn new(tid: String) -> Result<Self, &'static str> {
+        const RE_TID: OnceCell<Regex> = OnceCell::new();
+
+        if tid.len() != 13 {
+            Err("TID must be 13 characters")
+        } else if !RE_TID
+            .get_or_init(|| {
+                Regex::new(r"^[234567abcdefghij][234567abcdefghijklmnopqrstuvwxyz]{12}$").unwrap()
+            })
+            .is_match(&tid)
+        {
+            Err("Invalid TID")
+        } else {
+            Ok(Self(tid))
+        }
+    }
+
+    /// Returns the TID as a string slice.
+    pub fn as_str(&self) -> &str {
+        self.0.as_str()
+    }
+}
+
+/// A record key (`rkey`) used to name and reference an individual record within the same
+/// collection of an atproto repository.
+#[derive(Debug, Clone, PartialEq, Eq, serde::Serialize)]
+pub struct RecordKey(String);
+string_newtype!(RecordKey);
+
+impl RecordKey {
+    #[allow(
+        clippy::borrow_interior_mutable_const,
+        clippy::declare_interior_mutable_const
+    )]
+    /// Parses a `Record Key` from the given string.
+    pub fn new(s: String) -> Result<Self, &'static str> {
+        const RE_RKEY: OnceCell<Regex> = OnceCell::new();
+
+        if [".", ".."].contains(&s.as_str()) {
+            Err("Disallowed rkey")
+        } else if !RE_RKEY
+            .get_or_init(|| Regex::new(r"^[a-zA-Z0-9.\-_:~]{1,512}$").unwrap())
+            .is_match(&s)
+        {
+            Err("Invalid rkey")
+        } else {
+            Ok(Self(s))
+        }
+    }
+
+    /// Returns the record key as a string slice.
+    pub fn as_str(&self) -> &str {
+        self.0.as_str()
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use serde_json::{from_str, to_string};
@@ -422,7 +492,7 @@ mod tests {
     #[test]
     fn valid_datetime() {
         // From https://atproto.com/specs/lexicon#datetime
-        for valid in &[
+        for valid in [
             // preferred
             "1985-04-12T23:20:50.123Z",
             "1985-04-12T23:20:50.123456Z",
@@ -446,7 +516,7 @@ mod tests {
     #[test]
     fn invalid_datetime() {
         // From https://atproto.com/specs/lexicon#datetime
-        for invalid in &[
+        for invalid in [
             "1985-04-12",
             "1985-04-12T23:20Z",
             "1985-04-12T23:20:5Z",
@@ -488,7 +558,7 @@ mod tests {
     #[test]
     fn valid_did() {
         // From https://atproto.com/specs/did#examples
-        for valid in &[
+        for valid in [
             "did:plc:z72i7hdynmk6r22z27h6tvur",
             "did:web:blueskyweb.xyz",
             "did:method:val:two",
@@ -508,7 +578,7 @@ mod tests {
     #[test]
     fn invalid_did() {
         // From https://atproto.com/specs/did#examples
-        for invalid in &[
+        for invalid in [
             "did:METHOD:val",
             "did:m123:val",
             "DID:method:val",
@@ -528,7 +598,7 @@ mod tests {
     #[test]
     fn did_method() {
         // From https://atproto.com/specs/did#examples
-        for (method, did) in &[
+        for (method, did) in [
             ("did:plc", "did:plc:z72i7hdynmk6r22z27h6tvur"),
             ("did:web", "did:web:blueskyweb.xyz"),
             ("did:method", "did:method:val:two"),
@@ -540,14 +610,14 @@ mod tests {
                 "did:key:zQ3shZc2QzApp2oymGvQbzP8eKheVshBHbU4ZYjeXqwSKEn6N",
             ),
         ] {
-            assert_eq!(Did::new(did.to_string()).unwrap().method(), *method);
+            assert_eq!(Did::new(did.to_string()).unwrap().method(), method);
         }
     }
 
     #[test]
     fn valid_handle() {
         // From https://atproto.com/specs/handle#identifier-examples
-        for valid in &[
+        for valid in [
             "jay.bsky.social",
             "8.cn",
             "name.t--t", // not a real TLD, but syntax ok
@@ -573,7 +643,7 @@ mod tests {
     #[test]
     fn invalid_handle() {
         // From https://atproto.com/specs/handle#identifier-examples
-        for invalid in &[
+        for invalid in [
             "jo@hn.test",
             "💩.test",
             "john..test",
@@ -595,7 +665,7 @@ mod tests {
     #[test]
     fn valid_nsid() {
         // From https://atproto.com/specs/nsid#examples
-        for valid in &[
+        for valid in [
             "com.example.fooBar",
             "net.users.bob.ping",
             "a-0.b-1.c",
@@ -613,7 +683,7 @@ mod tests {
     #[test]
     fn invalid_nsid() {
         // From https://atproto.com/specs/nsid#examples
-        for invalid in &["com.exa💩ple.thing", "com.example"] {
+        for invalid in ["com.exa💩ple.thing", "com.example"] {
             assert!(
                 from_str::<Nsid>(&format!("\"{}\"", invalid)).is_err(),
                 "invalid NSID `{}` parsed as valid",
@@ -625,7 +695,7 @@ mod tests {
     #[test]
     fn nsid_parts() {
         // From https://atproto.com/specs/nsid#examples
-        for (nsid, domain_authority, name) in &[
+        for (nsid, domain_authority, name) in [
             ("com.example.fooBar", "com.example", "fooBar"),
             ("net.users.bob.ping", "net.users.bob", "ping"),
             ("a-0.b-1.c", "a-0.b-1", "c"),
@@ -633,15 +703,15 @@ mod tests {
             ("cn.8.lex.stuff", "cn.8.lex", "stuff"),
         ] {
             let nsid = Nsid::new(nsid.to_string()).unwrap();
-            assert_eq!(nsid.domain_authority(), *domain_authority);
-            assert_eq!(nsid.name(), *name);
+            assert_eq!(nsid.domain_authority(), domain_authority);
+            assert_eq!(nsid.name(), name);
         }
     }
 
     #[test]
     fn valid_language() {
         // From https://www.rfc-editor.org/rfc/rfc5646.html#appendix-A
-        for valid in &[
+        for valid in [
             // Simple language subtag:
             "de",         // German
             "fr",         // French
@@ -700,7 +770,7 @@ mod tests {
     #[test]
     fn invalid_language() {
         // From https://www.rfc-editor.org/rfc/rfc5646.html#appendix-A
-        for invalid in &[
+        for invalid in [
             "de-419-DE", // two region tags
             // use of a single-character subtag in primary position; note that there are a
             // few grandfathered tags that start with "i-" that are valid
@@ -709,6 +779,85 @@ mod tests {
             assert!(
                 from_str::<Language>(&format!("\"{}\"", invalid)).is_err(),
                 "invalid language `{}` parsed as valid",
+                invalid,
+            );
+        }
+    }
+
+    #[test]
+    fn valid_tid() {
+        for valid in ["3jzfcijpj2z2a", "7777777777777", "3zzzzzzzzzzzz"] {
+            assert!(
+                from_str::<Tid>(&format!("\"{}\"", valid)).is_ok(),
+                "valid TID `{}` parsed as invalid",
+                valid,
+            );
+        }
+    }
+
+    #[test]
+    fn invalid_tid() {
+        for invalid in [
+            // not base32
+            "3jzfcijpj2z21",
+            "0000000000000",
+            // too long/short
+            "3jzfcijpj2z2aa",
+            "3jzfcijpj2z2",
+            // old dashes syntax not actually supported (TTTT-TTT-TTTT-CC)
+            "3jzf-cij-pj2z-2a",
+            // high bit can't be high
+            "zzzzzzzzzzzzz",
+            "kjzfcijpj2z2a",
+        ] {
+            assert!(
+                from_str::<Tid>(&format!("\"{}\"", invalid)).is_err(),
+                "invalid TID `{}` parsed as valid",
+                invalid,
+            );
+        }
+    }
+
+    #[test]
+    fn valid_rkey() {
+        // From https://atproto.com/specs/record-key#examples
+        for valid in [
+            "3jui7kd54zh2y",
+            "self",
+            "literal:self",
+            "example.com",
+            "~1.2-3_",
+            "dHJ1ZQ",
+            "pre:fix",
+            "_",
+        ] {
+            assert!(
+                from_str::<RecordKey>(&format!("\"{}\"", valid)).is_ok(),
+                "valid rkey `{}` parsed as invalid",
+                valid,
+            );
+        }
+    }
+
+    #[test]
+    fn invalid_rkey() {
+        // From https://atproto.com/specs/record-key#examples
+        for invalid in [
+            "alpha/beta",
+            ".",
+            "..",
+            "#extra",
+            "@handle",
+            "any space",
+            "any+space",
+            "number[3]",
+            "number(3)",
+            "\"quote\"",
+            "dHJ1ZQ==",
+        ] {
+            assert!(
+                from_str::<RecordKey>(&format!("\"{}\"", invalid)).is_err(),
+                "invalid rkey `{}` parsed as valid",
                 invalid,
             );
         }