+++ atrium-api/CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- `atrium_api::types::string::Tid`
+- `atrium_api::types::string::RecordKey`
+  - moved from `atrium_api::types::RecordKey`
+
+### Removed
+- `atrium_api::types::RecordKey`
+  - moved to `atrium_api::types::string::RecordKey`
+
 ## [0.20.1](https://github.com/sugyan/atrium/compare/atrium-api-v0.20.0...atrium-api-v0.20.1) - 2024-04-17
 
 ### Added